Major changes in the optitrack calibration
- Changes the torso marker to be fit on the shoulder_center instead of
the spine frame. This has required to also change the namespace
"/optitrack" to "/opt" as the marker name
"/optitrack/human/shoulder_center" is too long for motive to handle.

- Add a verification on the initial base frame guess for the
optimization such that x base axis always face the robot. This handles
cases were the initial guess was facing opposite which lead to not
converging calibration